requirements.txt
asgiref==3.4.1
Django==3.2.7
dnspython==1.16.0
pymongo==3.12.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.2
urllib3==1.26.6
uuid==1.30
